requirements.txt
# FastAPI web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Computer vision and image processing
opencv-python>=4.8.0
numpy>=1.24.0

# System utilities
psutil>=5.9.0
pyudev>=0.24.0

# EXIF tool for image metadata
pyexiftool>=0.5.0
